--- lib/core/constants/app_size.dart
@@ -1,60 +1,0 @@
-class AppMargin {
-  static const double m8 = 8.0;
-  static const double m12 = 12.0;
-  static const double m14 = 14.0;
-  static const double m16 = 16.0;
-  static const double m18 = 18.0;
-  static const double m20 = 20.0;
-  static const double m24 = 24.0;
-}
-
-class AppPadding {
-  static const double p1 = 1.0;
-  static const double p4 = 4.0;
-  static const double p7 = 7.0;
-  static const double p8 = 8.0;
-  static const double p10 = 10.0;
-  static const double p12 = 12.0;
-  static const double p14 = 14.0;
-  static const double p16 = 16.0;
-  static const double p18 = 18.0;
-  static const double p20 = 20.0;
-  static const double p24 = 24.0;
-}
-
-class AppSize {
-  static const double s0 = 0.0;
-  static const double s0_012 = 0.012;
-  static const double s0_013 = 0.013;
-  static const double s0_015 = 0.015;
-  static const double s0_03 = 0.03;
-  static const double s0_9 = 0.9;
-  static const double s1 = 1.0;
-  static const double s1_5 = 1.5;
-  static const double s2 = 2.0;
-  static const double s3 = 3.0;
-  static const double s4 = 4.0;
-  static const double s5 = 5.0;
-  static const double s8 = 8.0;
-  static const double s10 = 10.0;
-  static const double s12 = 12.0;
-  static const double s14 = 14.0;
-  static const double s16 = 16.0;
-  static const double s18 = 18.0;
-  static const double s20 = 20.0;
-  static const double s22 = 22.0;
-  static const double s24 = 24.0;
-  static const double s26 = 26.0;
-  static const double s30 = 30.0;
-  static const double s35 = 35.0;
-  static const double s40 = 40.0;
-  static const double s50 = 50.0;
-  static const double s60 = 60.0;
-  static const double s65 = 65.0;
-  static const double s70 = 70.0;
-  static const double s75 = 75.0;
-  static const double s100 = 100.0;
-  static const double s180 = 180.0;
-  static const double s200 = 200.0;
-  static const double s250 = 250.0;
-}
