--- lib/core/constants/icon_broken.dart
@@ -1,108 +1,0 @@
-import 'package:flutter/widgets.dart';
-
-class IconBroken {
-  IconBroken._();
-
-  static const String _fontFamily = 'IconBroken';
-
-  static const IconData User = IconData(0xe900, fontFamily: _fontFamily);
-  static const IconData User1 = IconData(0xe901, fontFamily: _fontFamily);
-  static const IconData Activity = IconData(0xe902, fontFamily: _fontFamily);
-  static const IconData Add_User = IconData(0xe903, fontFamily: _fontFamily);
-  static const IconData Arrow___Down_2 = IconData(0xe904, fontFamily: _fontFamily);
-  static const IconData Arrow___Down_3 = IconData(0xe905, fontFamily: _fontFamily);
-  static const IconData Arrow___Down_Circle = IconData(0xe906, fontFamily: _fontFamily);
-  static const IconData Arrow___Down_Square = IconData(0xe907, fontFamily: _fontFamily);
-  static const IconData Arrow___Down = IconData(0xe908, fontFamily: _fontFamily);
-  static const IconData Arrow___Left_2 = IconData(0xe909, fontFamily: _fontFamily);
-  static const IconData Arrow___Left_3 = IconData(0xe90a, fontFamily: _fontFamily);
-  static const IconData Arrow___Left_Circle = IconData(0xe90b, fontFamily: _fontFamily);
-  static const IconData Arrow___Left_Square = IconData(0xe90c, fontFamily: _fontFamily);
-  static const IconData Arrow___Left = IconData(0xe90d, fontFamily: _fontFamily);
-  static const IconData Arrow___Right_2 = IconData(0xe90e, fontFamily: _fontFamily);
-  static const IconData Arrow___Right_3 = IconData(0xe90f, fontFamily: _fontFamily);
-  static const IconData Arrow___Right_Circle = IconData(0xe910, fontFamily: _fontFamily);
-  static const IconData Arrow___Right_Square = IconData(0xe911, fontFamily: _fontFamily);
-  static const IconData Arrow___Right = IconData(0xe912, fontFamily: _fontFamily);
-  static const IconData Arrow___Up_2 = IconData(0xe913, fontFamily: _fontFamily);
-  static const IconData Arrow___Up_3 = IconData(0xe914, fontFamily: _fontFamily);
-  static const IconData Arrow___Up_Circle = IconData(0xe915, fontFamily: _fontFamily);
-  static const IconData Arrow___Up_Square = IconData(0xe916, fontFamily: _fontFamily);
-  static const IconData Arrow___Up = IconData(0xe917, fontFamily: _fontFamily);
-  static const IconData Bag_2 = IconData(0xe918, fontFamily: _fontFamily);
-  static const IconData Bag = IconData(0xe919, fontFamily: _fontFamily);
-  static const IconData Bookmark = IconData(0xe91a, fontFamily: _fontFamily);
-  static const IconData Buy = IconData(0xe91b, fontFamily: _fontFamily);
-  static const IconData Calendar = IconData(0xe91c, fontFamily: _fontFamily);
-  static const IconData Call_Missed = IconData(0xe91d, fontFamily: _fontFamily);
-  static const IconData Call_Silent = IconData(0xe91e, fontFamily: _fontFamily);
-  static const IconData Call = IconData(0xe91f, fontFamily: _fontFamily);
-  static const IconData Calling = IconData(0xe920, fontFamily: _fontFamily);
-  static const IconData Camera = IconData(0xe921, fontFamily: _fontFamily);
-  static const IconData Category = IconData(0xe922, fontFamily: _fontFamily);
-  static const IconData Chart = IconData(0xe923, fontFamily: _fontFamily);
-  static const IconData Chat = IconData(0xe924, fontFamily: _fontFamily);
-  static const IconData Close_Square = IconData(0xe925, fontFamily: _fontFamily);
-  static const IconData Danger = IconData(0xe926, fontFamily: _fontFamily);
-  static const IconData Delete = IconData(0xe927, fontFamily: _fontFamily);
-  static const IconData Discount = IconData(0xe928, fontFamily: _fontFamily);
-  static const IconData Discovery = IconData(0xe929, fontFamily: _fontFamily);
-  static const IconData Document = IconData(0xe92a, fontFamily: _fontFamily);
-  static const IconData Download = IconData(0xe92b, fontFamily: _fontFamily);
-  static const IconData Edit_Square = IconData(0xe92c, fontFamily: _fontFamily);
-  static const IconData Edit = IconData(0xe92d, fontFamily: _fontFamily);
-  static const IconData Filter_2 = IconData(0xe92e, fontFamily: _fontFamily);
-  static const IconData Filter = IconData(0xe92f, fontFamily: _fontFamily);
-  static const IconData Folder = IconData(0xe930, fontFamily: _fontFamily);
-  static const IconData Game = IconData(0xe931, fontFamily: _fontFamily);
-  static const IconData Graph = IconData(0xe932, fontFamily: _fontFamily);
-  static const IconData Heart = IconData(0xe933, fontFamily: _fontFamily);
-  static const IconData Hide = IconData(0xe934, fontFamily: _fontFamily);
-  static const IconData Home = IconData(0xe935, fontFamily: _fontFamily);
-  static const IconData Image_2 = IconData(0xe936, fontFamily: _fontFamily);
-  static const IconData Image = IconData(0xe937, fontFamily: _fontFamily);
-  static const IconData Info_Circle = IconData(0xe938, fontFamily: _fontFamily);
-  static const IconData Info_Square = IconData(0xe939, fontFamily: _fontFamily);
-  static const IconData Location = IconData(0xe93a, fontFamily: _fontFamily);
-  static const IconData Lock = IconData(0xe93b, fontFamily: _fontFamily);
-  static const IconData Login = IconData(0xe93c, fontFamily: _fontFamily);
-  static const IconData Logout = IconData(0xe93d, fontFamily: _fontFamily);
-  static const IconData Message = IconData(0xe93e, fontFamily: _fontFamily);
-  static const IconData More_Circle = IconData(0xe93f, fontFamily: _fontFamily);
-  static const IconData More_Square = IconData(0xe940, fontFamily: _fontFamily);
-  static const IconData Notification = IconData(0xe941, fontFamily: _fontFamily);
-  static const IconData Paper_Download = IconData(0xe942, fontFamily: _fontFamily);
-  static const IconData Paper_Fail = IconData(0xe943, fontFamily: _fontFamily);
-  static const IconData Paper_Negative = IconData(0xe944, fontFamily: _fontFamily);
-  static const IconData Paper_Plus = IconData(0xe945, fontFamily: _fontFamily);
-  static const IconData Paper_Upload = IconData(0xe946, fontFamily: _fontFamily);
-  static const IconData Paper = IconData(0xe947, fontFamily: _fontFamily);
-  static const IconData Password = IconData(0xe948, fontFamily: _fontFamily);
-  static const IconData Play = IconData(0xe949, fontFamily: _fontFamily);
-  static const IconData Plus = IconData(0xe94a, fontFamily: _fontFamily);
-  static const IconData Profile = IconData(0xe94b, fontFamily: _fontFamily);
-  static const IconData Scan = IconData(0xe94c, fontFamily: _fontFamily);
-  static const IconData Search = IconData(0xe94d, fontFamily: _fontFamily);
-  static const IconData Send = IconData(0xe94e, fontFamily: _fontFamily);
-  static const IconData Setting = IconData(0xe94f, fontFamily: _fontFamily);
-  static const IconData Shield_Done = IconData(0xe950, fontFamily: _fontFamily);
-  static const IconData Shield_Fail = IconData(0xe951, fontFamily: _fontFamily);
-  static const IconData Show = IconData(0xe952, fontFamily: _fontFamily);
-  static const IconData Star = IconData(0xe953, fontFamily: _fontFamily);
-  static const IconData Swap = IconData(0xe954, fontFamily: _fontFamily);
-  static const IconData Tick_Square = IconData(0xe955, fontFamily: _fontFamily);
-  static const IconData Ticket_Star = IconData(0xe956, fontFamily: _fontFamily);
-  static const IconData Ticket = IconData(0xe957, fontFamily: _fontFamily);
-  static const IconData Time_Circle = IconData(0xe958, fontFamily: _fontFamily);
-  static const IconData Time_Square = IconData(0xe959, fontFamily: _fontFamily);
-  static const IconData Unlock = IconData(0xe95a, fontFamily: _fontFamily);
-  static const IconData Upload = IconData(0xe95b, fontFamily: _fontFamily);
-  static const IconData Video = IconData(0xe95c, fontFamily: _fontFamily);
-  static const IconData Voice_2 = IconData(0xe95d, fontFamily: _fontFamily);
-  static const IconData Voice = IconData(0xe95e, fontFamily: _fontFamily);
-  static const IconData Volume_Down = IconData(0xe95f, fontFamily: _fontFamily);
-  static const IconData Volume_Off = IconData(0xe960, fontFamily: _fontFamily);
-  static const IconData Volume_Up = IconData(0xe961, fontFamily: _fontFamily);
-  static const IconData Wallet = IconData(0xe962, fontFamily: _fontFamily);
-  static const IconData Work = IconData(0xe963, fontFamily: _fontFamily);
-}
