--- lib/core/themes/colors_manager.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ColorManager {
-  static const blue = Color(0xFF285DC2);
-  static const purple = Color(0xFF4B36C1);
-  static const lightGreen = Color(0xFF73D197);
-  static const pink = Color(0xFFE7647F);
-  static const yellow = Color(0xFFF9CF67);
-  static const orange = Color(0xFFF39970);
-  static const greyColor = Color(0xff323232);
-  static const lightGrey = Color(0xFFE8EAF0);
-  static const lightGrey2 = Color(0xFFA5A6AC);
-  static const error = Color(0xFFfa0000);
-}
